Rendering the enemy!
The position (enemyStart and enemyStop) are a bit off (he's floating in midair) but he's rendering properly!

+++ src/Processing/SuperCold/Enemy.pde
@@ -9,7 +9,7 @@ class Enemy {
   int movementTimer;
   int dyingTimer;
   char dyingStage;
-  int enemyWidth = 3;
+  float enemyWidth = 3;
   float enemyHeight = 0;
   PImage texture;
